server/proxy: Move client input callbacks to pf_input
